pyproject.toml
[tool.poetry]
name = "optimal-transport-introduction"
version = "0.1.0"
description = ""
authors = ["kohei <koheitech001@gmail.com>"]
license = "Apach-2.0"
readme = "README.md"
packages = [{include = "optimal_transport_introduction"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
matplotlib = "^3.7.1"
tensorflow-cpu = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
